backend/requirements.txt
fastapi
supabase
python-dotenv
python-multipart
uvicorn
bcrypt
sqlalchemy
psycopg2-binary
pyjwt
pydantic[email]